ES6 and multiple shuffled decks
Updated the codebase to utilise ES6 JavaScript, enhancing readability, maintainability, and taking advantage of modern language features.

Altered the file structure to include a src and test folder.

Implemented a new function to replicate the deck n times, enabling the concept of 'multiple decks'. The numberOfDecks variable is declared so that the desired number of decks can easily be changed.

Added a function for shuffling the decks to introduce randomness to the cards. (Ensured that shuffling occurs after deck replication to prevent repetition of the same shuffled decks.)

Added a summary in the readme.md file to summarise what the project is about.

In the test.js file, incorporated a loop to visually represent the randomness of cards outputted.

Additionally, a loop was included to demonstrate the functionality of the slice method, illustrating its ability to remove a card on each iteration until the entire list has been covered.

Comments were inserted throughout the codebase to aid others.